requirements.txt
# Core dependencies - Python 3.10 compatible
opencv-python>=4.5.0,<5.0.0
numpy>=1.21.0,<1.25.0
scipy>=1.7.0,<1.12.0
scikit-learn>=1.0.0,<1.4.0
pandas>=1.3.0,<2.1.0

# Deep learning frameworks - Python 3.10 compatible
torch>=1.12.0,<2.1.0
torchvision>=0.13.0,<0.16.0

# Computer vision and face analysis
# MediaPipe - using version compatible with Python 3.10
mediapipe>=0.8.10,<0.11.0
# dlib - optional, can be installed separately if needed
# dlib>=19.22.0
imutils>=0.5.4

# Face detection alternative (if dlib not available)
opencv-contrib-python>=4.5.0,<5.0.0

# Image processing
Pillow>=9.0.0,<11.0.0
imageio>=2.15.0,<3.0.0

# Data handling and visualization
matplotlib>=3.5.0,<3.8.0
seaborn>=0.11.0,<0.13.0
tqdm>=4.60.0,<5.0.0

# Configuration and utilities
PyYAML>=6.0,<7.0
# configparser is built-in to Python 3.10
# argparse is built-in to Python 3.10

# Dataset handling
h5py>=3.6.0,<4.0.0

# Optional GUI support
# PyQt5>=5.15.0  # Uncomment if GUI needed

# Video processing
# ffmpeg-python>=0.2.0  # Optional

# Additional utilities
pathlib2>=2.3.0; python_version < "3.4"

